+++ src/main/java/app/service/CalculateBOM.java
@@ -198,7 +198,7 @@ public class CalculateBOM {
     public int calculateTotalPriceFromBOM() {
         int total = 0;
         for (BOM bom : bomList) {
-            double pricePerMeter = bom.getProductVariant().getProduct().getPrice();
+            int pricePerMeter = bom.getProductVariant().getProduct().getPrice();
             int quantity = bom.getQuantity();
             total += pricePerMeter * quantity;
         }
